Prompt to add user to fuse group on RHEL 6
RHEL 6 requires that users be in the fuse group to mount FUSE filesystems.
Add an autoactivated D-Bus service which can make the necessary changes
to group membership, and call it when vmnetx is launched.  If changes
are necessary, use polkit to prompt the user for an administrator
password.  Once added to the group, reexecute vmnetx through "sg" to
update the supplementary group list so that the user doesn't have to log
out and back in.

Closes #10.