Fix incorrect sizing of VMWindow
We were firing the viewer-resize signal before the SpiceClientGtk.Display
had a chance to change its size request.  This caused a race with the
window resize that we force, sometimes resulting in an incorrect window
size and a downscaled guest.  Fix this by firing viewer-resize only after
the Display updates its size request.