generate: Remove backward-incompatible defaults from domain XML
libvirt rewrites domain XML for newly-defined domains before the XML hits
the disk.  vmnetx-generate -a therefore writes XML that does not match
our schema.  So, when reading domain XML for vmnetx-generate or
vmnetx-generate -c, apply a series of transforms that remove backward-
incompatible default settings.  The transforms will need to be updated
for future libvirt changes, but without them, manual curator intervention
in the domain XML would be required for every new VM.