vmnetx.controller.local: Recompress memory with LZO on first launch
Memory images are typically compressed with LZMA, which causes launch
to be CPU-bound in decompression if the memory image is locally cached.
30 seconds after finishing VM launch, start a background thread to
recompress the memory image with LZO if not already recompressed.  On
subsequent launches, resume from the LZO-compressed image.

We could, instead, recompress while qemu is still reading the memory
image.  When fetching the memory image over a slow link, this would be a
good use of unused CPU time.  However, it would be difficult to ensure
that this never slows the launch (e.g. on single-core systems, fast
networks, or for memory images already in cache).  We would also need
to ensure recompression never gets ahead of qemu in reading the memory
image, since that would cause misleading progress bar updates.  So we
stick with the simple solution for now.

Fixes #52.